+++ src/components/Button.tsx
@@ -1,5 +1,5 @@
 import React, { FC } from 'react';
-import { gameStatusWord } from '../App'
+import { gameStatusWord, GameStatus } from '../constants'
 
 type Props = {
   status: gameStatusWord
@@ -11,10 +11,10 @@ type Props = {
 const Button :FC<Props> = ({ status, onStart, onRestart, onStop }) => {
   return (
     <div className="button">
-      { status === "gameover" && <button className="btn btn-gameover" onClick={ onRestart }>gameover</button> }
-      { status === "init" && <button className="btn btn-init" onClick={ onStart }>start</button> }
-      { status === "suspended" && <button className="btn btn-suspended" onClick={ onStart }>start</button> }
-      { status === "playing" && <button className="btn btn-playing" onClick={ onStop }>stop</button> }
+       { status === GameStatus.gameover && <button className="btn btn-gameover" onClick={onRestart}>gameover</button> }
+       { status === GameStatus.init && <button className="btn btn-init" onClick={onStart}>start</button> }
+       { status === GameStatus.suspended && <button className="btn btn-suspended" onClick={onStart}>start</button> }
+       { status === GameStatus.playing && <button className="btn btn-playing" onClick={onStop}>stop</button> }
     </div>
   );
 };
