+++ src/utils/index.ts
@@ -1,4 +1,5 @@
-import { initialPosi } from '../App';
+import { initialPosi } from '../constants';
+import { unsubscribe } from '../hooks/useSnakGame';
 
 export const getFoodPosition = (fieldSize: number, excludes: initialPosi[]) => {
   while(true) {
@@ -23,4 +24,19 @@ export const initFields = (fieldSize: number, snake: initialPosi) => {
   fields[food.y][food.x] = 'food';
 
   return fields;
+};
+
+export const isCollision = (fieldSize: number, position: initialPosi) => {
+  if(position.y < 0 || position.x < 0) {
+    return true;
+  };
+  if(position.y > fieldSize -1 || position.x > fieldSize -1) {
+    unsubscribe()
+    return true;
+  };
+  return false;
+};
+
+export const isEatingMysellf = (fields: string[][], position: initialPosi) => {
+  return fields[position.y][position.x] === 'snake';
 };
